+++ src/App.jsx
@@ -10,12 +10,12 @@ import { useDispatch } from 'react-redux';
 
 import styled from '@emotion/styled';
 
-import HomePage from './HomePage';
-import AboutPage from './AboutPage';
-import LoginPage from './LoginPage';
-import RestaurantsPage from './RestaurantsPage';
-import RestaurantPage from './RestaurantPage';
-import NotFoundPage from './NotFoundPage';
+import HomePage from './pages/HomePage';
+import AboutPage from './pages/AboutPage';
+import LoginPage from './pages/LoginPage';
+import RestaurantsPage from './pages/RestaurantsPage';
+import RestaurantPage from './pages/RestaurantPage';
+import NotFoundPage from './pages/NotFoundPage';
 
 import { setAccessToken } from './slice';
 
@@ -23,7 +23,7 @@ import { loadItem } from './services/storage';
 
 import {
   lightGray, darkBlue, skyBlue, pastelPink,
-} from './styles/constants';
+} from './assets/styles/colors';
 
 const Header = styled.header({
   textAlign: 'center',
