+++ src/assets/documents/about-eatgo.js
@@ -1,0 +1,7 @@
+const aboutEatgo = {
+  body: `이 서비스는 영국에서 시작되었습니다.
+
+이 서비스를 보셨다면 주변에 있는 20명에게 추천하셔야 합니다.`,
+};
+
+export default aboutEatgo;
