+++ src/assets/styles/colors.js
@@ -1,0 +1,4 @@
+export const lightGray = '#f5f5f5';
+export const darkBlue = '#364f6b';
+export const skyBlue = '#3fc1c9';
+export const pastelPink = '#fc5185';
