+++ src/components/ContentWrapper.jsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { lightGray, darkBlue } from './styles/constants';
+import { lightGray, darkBlue } from '../assets/styles/colors';
 
 const ContentWrapper = styled.div({
   borderStyle: 'solid none',
@@ -9,6 +9,7 @@ const ContentWrapper = styled.div({
   padding: '20px 20px',
   margin: '30px auto',
   backgroundColor: lightGray,
+  whiteSpace: 'pre-wrap',
   fontSize: '1.2em',
   '& h3': {
     color: darkBlue,
