+++ src/components/login/LoginFormContainer.jsx
@@ -9,9 +9,9 @@ import {
   changeLoginField,
   requestLogin,
   logout,
-} from './slice';
+} from '../../slice';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
