+++ src/components/login/LogoutForm.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from './Button';
+import Button from '../Button';
 
 export default function LogoutForm({ onClick }) {
   return (
