+++ src/components/restaurants/CategoriesContainer.test.jsx
