+++ src/components/restaurants/RegionsContainer.jsx
@@ -2,16 +2,16 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import OptionList from './OptionList';
-import OptionItem from './OptionItem';
-import Button from './Button';
+import OptionList from '../OptionList';
+import OptionItem from '../OptionItem';
+import Button from '../Button';
 
 import {
   selectRegion,
   loadRestaurants,
-} from './slice';
+} from '../../slice';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
