+++ src/components/restaurants/RestaurantsContainer.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 import { useSelector } from 'react-redux';
 
-import MenuList from './MenuList';
-import MenuItem from './MenuItem';
+import MenuList from '../MenuList';
+import MenuItem from '../MenuItem';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function RestaurantsContainer({ onClickRestaurant }) {
   const restaurants = useSelector(get('restaurants'));
