+++ src/pages/LoginPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import LoginFormContainer from './LoginFormContainer';
-import PageWrapper from './PageWrapper';
+import LoginFormContainer from '../components/login/LoginFormContainer';
+import PageWrapper from '../components/PageWrapper';
 
 export default function LoginPage() {
   return (
