+++ src/pages/NotFoundPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PageWrapper from './PageWrapper';
+import PageWrapper from '../components/PageWrapper';
 
 export default function NotFoundPage() {
   return (
