+++ src/pages/RestaurantPage.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { useParams } from 'react-router-dom';
 
-import RestaurantContainer from './RestaurantContainer';
-import PageWrapper from './PageWrapper';
+import RestaurantContainer from '../components/restaurant/RestaurantContainer';
+import PageWrapper from '../components/PageWrapper';
 
 export default function RestaurantPage({ params }) {
   const { id } = params || useParams();
