+++ src/slice.js
@@ -10,7 +10,7 @@ import {
 } from './services/api';
 
 import { saveItem } from './services/storage';
-import { equal } from './utils';
+import { equal } from './utils/utils';
 
 const { actions, reducer } = createSlice({
   name: 'application',
